+++ src/main/java/annotation/RequestMapping.java
@@ -1,6 +1,6 @@
 package annotation;
 
-import controller.HttpMethod;
+import support.HttpMethod;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
