+++ src/main/java/support/HttpMethod.java
@@ -1,4 +1,4 @@
-package controller;
+package support;
 
 public enum HttpMethod {
     GET
